Delete strange cast.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@24012 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/tra.cpp
@@ -2414,7 +2414,7 @@ static tx_inv_page* fetch_inventory_page(
  **************************************/
 	SET_TDBB(tdbb);
 
-	window->win_page = inventory_page(tdbb, (int) sequence);
+	window->win_page = inventory_page(tdbb, sequence);
 	tx_inv_page* tip =
 		(tx_inv_page*) CCH_FETCH(tdbb, window, lock_level, pag_transactions);
